# eval: performance improvements